cpu, x86: Keep the TLB warm up before CPU switch
Previously, before a CPU was switched out, the TLB was always flushed
Now, we first call takeOverFrom with the TLB. We only flush
the TLB right before the CPU is switched in.
This changeset also contains the needed code for x86 to
takeOverFrom with the TLB, similar changes may be needed for the ARM
architecture.
With this changeset, when you switch from atomic to timing mode the
TLB is warm.